+++ src/main/java/roomescape/member/WebConfiguration.java
@@ -1,0 +1,21 @@
+package roomescape.member;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class WebConfiguration implements WebMvcConfigurer {
+    private LoginMemberArgumentResolver loginMemberArgumentResolver;
+
+    public WebConfiguration(LoginMemberArgumentResolver loginMemberArgumentResolver) {
+        this.loginMemberArgumentResolver = loginMemberArgumentResolver;
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(loginMemberArgumentResolver);
+    }
+}
